# Conflicts:
#	src/structures/IRoom.java
#	src/structures/data/DataGame.java
#	src/structures/data/DataRoom.java
#	src/structures/data/DataView.java

+++ src/structures/IRoom.java
@@ -3,22 +3,23 @@ package structures;
 import structures.data.DataView;
 
 public interface IRoom {
-	
+
 	void addObject(IObject o);
-	
+
 	void removeObject(IObject o);
-	
-	void setSize(double width, double height);
-	
+
+	void setSize(int width, int height);
+
 	double[] getSize();
-	
+
 	void setBackground(IObject o);
-	
+
 	Object getBackground();
-	
+
 	void setView(DataView view);
-	
+
 	DataView getView();
 
 	String getName();
+
 }
